[image] fix: missing return in plt error state
[model] fix: danglymesh constraints are float..
[model] fix: wrong number of bones

+++ lib/nw/model/Mdl.hpp
@@ -448,7 +448,7 @@ struct AnimeshNode : public TrimeshNode {
 struct DanglymeshNode : public TrimeshNode {
     DanglymeshNode(std::string name_);
 
-    std::vector<uint32_t> constraints;
+    std::vector<float> constraints;
     float displacement;
     float period;
     float tightness;
